chore: separate test file
